setup: use kubevirt platform
We need the kubevirt platform and not `None` because we need the nodes
to be created by the kubevirt provider and not by kcli.

This is important later on for the hypershift operator to track after the machines
that created for the nodes. if kcli is the component that creates the VMs (machines)
then the operator doesn't track them and does no apply configuration changes to
the machines when new machineConfig or kubeletConfig applied.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>